reasoning_bank/requirements.txt
# ReasoningBank 依赖

# LLM API
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# Embedding
sentence-transformers>=2.2.0
torch>=2.0.0

# 数据处理
numpy>=1.24.0
pyyaml>=6.0

# 异步支持
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# 日志和工具
tqdm>=4.65.0
colorama>=0.4.6

# 可选：多轮环境依赖（如需运行 ALFWorld/ScienceWorld 请取消注释）
# alfworld>=0.3.0
# scienceworld>=1.1.0
# textworld>=1.5.0

